# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "一个基于 Gin + GORM + Redis + RabbitMQ 的在线书店后端服务，提供用户认证、书籍管理、购物车、订单及秒杀等功能。",
        "title": "Project1_Shop API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Me",
            "email": "frottenice@outlook.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:9090",
    "basePath": "/",
    "paths": {
...
        "/api/cart": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "获取当前登录用户的购物车中所有商品（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "获取购物车列表",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "更新当前登录用户购物车中指定书籍的数量（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "更新购物车中书籍数量",
                "parameters": [
                    {
                        "description": "更新参数（book_id 和 quantity 必填）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "登录用户将书籍加入购物车，若未指定数量则默认为 1（需要登录）",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "添加书籍到购物车",
                "parameters": [
                    {
                        "description": "购物车参数（book_id 必填，quantity 可选）",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CartParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "清空当前登录用户购物车中的所有商品（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "清空购物车",
                "responses": {
                    "200": {
                        "description": "未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
        "/api/cart/{book_id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "从当前登录用户的购物车中移除指定书籍（需要登录）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "购物车"
                ],
                "summary": "删除购物车中指定书籍",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "要删除的书籍 ID",
                        "name": "book_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "参数错误 / 未登录 / 服务器繁忙",
                        "schema": {
                            "$ref": "#/definitions/models.ResponseData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.CartParam": {
            "type": "object",
            "required": [
                "book_id"
            ],
            "properties": {
                "book_id": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "models.ResponseData": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {}
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "格式：Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}